+++ src/components/NavBar/Navbar.js
@@ -1,11 +1,20 @@
 import React, { useState } from 'react';
 import { NavLink } from 'react-router-dom';  // Use NavLink instead of Link
-import '../NavBar/Navbar.css';
+import { Link, useNavigate, useLocation } from 'react-router-dom';
+import './Navbar.css';
 import collegeLogo from '../../Assets/CollegeLogo.png';
 import 'font-awesome/css/font-awesome.min.css';
 
 const Navbar = () => {
   const [isMobile, setIsMobile] = useState(false);
+  const location = useLocation();
+  const navigate = useNavigate(); // To programmatically navigate
+
+  const isActive = (path) => location.pathname === path;
+
+  const handleLoginClick = () => {
+    navigate("/login"); // Navigate to the login page
+  };
 
   return (
     <nav className="navbar">
@@ -95,7 +104,14 @@ const Navbar = () => {
             Discussion
           </NavLink>
         </li>
-        <li><button>Login</button></li>
+        <li>
+          <button
+            className={isActive("/login") ? "active" : ""}
+            onClick={handleLoginClick}
+          >
+            Login
+          </button>
+        </li>
       </ul>
       <button className="mobile-menu-icon" onClick={() => setIsMobile(!isMobile)}>
         {isMobile ? <i className="fas fa-times"></i> : <i className="fas fa-bars"></i>}
