--- src/components/Navbar.css
@@ -1,325 +1,0 @@
-/* Resetting default margin and padding */
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-/* Set up the body */
-body {
-  font-family: 'Roboto', sans-serif;
-  background-color: #0f0f1c;
-  color: white;
-  margin: 0;
-  padding: 0;
-  overflow-x: hidden;
-}
-
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 5px 10px;
-  background: linear-gradient(to right, #000 25%, #1b2251, #21035a);
-  position: sticky;
-  top: 0;
-  z-index: 1;
-  width: 100%;
-  box-shadow: 0 15px 40px rgba(0, 0, 0, 0.6);
-  border-bottom: 2px solid rgba(255, 255, 255, 0.2);
-  backdrop-filter: blur(10px);
-  transition: all 0.3s ease;
-  animation: backgroundAnimation 5s infinite alternate;
-}
-
-nav {
-  animation: backgroundAnimation 5s infinite alternate;
-  background-size: 400% 400%;
-}
-
-.navbar-logo {
-  display: flex;
-  align-items: center;
-  font-size: 2rem;
-  color: #fff;
-  font-weight: bold;
-  letter-spacing: 3px;
-  text-transform: uppercase;
-  /*text-shadow: 0 0 8px rgba(15, 102, 102, 0.6);*/
-}
-
-.navbar-logo img {
-  height: 50px;
-  margin-right: 15px;
-  animation: logoGlow 1.5s infinite alternate;
-}
-/*
-@keyframes logoGlow {
-  0% {
-    box-shadow: 0 0 20px rgba(2, 21, 21, 0.9), 0 0 40px rgba(0, 255, 255, 0.6);
-  }
-  100% {
-    box-shadow: 0 0 40px rgba(1, 7, 7, 0.9), 0 0 60px rgba(0, 255, 255, 0.6);
-  }
-}
-*/
-.navbar-links {
-  display: flex;
-  gap: 30px;
-}
-
-.navbar-links li {
-  list-style-type: none;
-}
-
-.navbar-links a {
-  font-family:serif;
-  color: #f1f1f1;
-  text-decoration: none;
-  font-size: 18px;
-  font-weight: 500;
-  padding: 10px;
-  border-radius: 5px;
-  text-transform: uppercase;
-  transition: 0.4s;
-  position: relative;
-  z-index: 2;
-  top: 10px;
-}
-
-.navbar-links a:hover {
-  color: #0ff;
-  transform: translateY(-5px);
-  text-shadow: 0 0 10px rgba(0, 255, 255, 1), 0 0 20px rgba(0, 255, 255, 0.8);
-}
-
-.navbar-links a::before {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  /*background: rgba(0, 255, 255, 0.2);*/
-  transition: all 0.4s;
-  border-radius: 5px;
-  z-index: -1;
-  margin-top: auto;
-}
-
-.navbar-links li button {
-  background-color: transparent;
-  border: 2px solid #00ffcc; /* Futuristic neon border */
-  color: #00ffcc;
-  font-size: 1rem;
-  padding: 10px 20px;
-  border-radius: 5px;
-  cursor: pointer;
-  position: relative;
-  z-index: 1;
-  transition: all 0.4s ease;
-  overflow: hidden;
-}
-
-.navbar-links li button::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: -100%;
-  width: 100%;
-  height: 100%;
-  background: linear-gradient(120deg, rgba(0, 255, 204, 0.3), rgba(0, 255, 204, 0.1));
-  transition: 0.5s ease;
-  z-index: -1;
-}
-
-.navbar-links li button:hover::before {
-  left: 100%;
-}
-
-.navbar-links li button:hover {
-  color: #fff;
-  border-color: #fff;
-}
-
-.navbar-links li button:focus {
-  outline: none;
-}
-
-.navbar-links li button:active {
-  transform: scale(0.95);
-}
-
-/* Additional futuristic glow effect */
-.navbar-links li button::after {
-  content: "";
-  position: absolute;
-  top: -5px;
-  left: -5px;
-  right: -5px;
-  bottom: -5px;
-  background: rgba(0, 255, 204, 0.2);
-  border-radius: 5px;
-  z-index: -1;
-  opacity: 0;
-  transition: opacity 0.4s ease;
-}
-
-.navbar-links li button:hover::after {
-  opacity: 1;
-}
-
-.navbar-links a:hover::before {
-  width: 0%;
-  height: 0%;
-  transform: scale(5);
-}
-
-/* Mobile Menu Icon */
-.mobile-menu-icon {
-  display: none;
-  background: none;
-  border: none;
-  color: white;
-  font-size: 30px;
-  cursor: pointer;
-  transition: 0.3s ease;
-}
-
-.mobile-menu-icon i {
-  color:#fff;
-}
-
-/* Mobile Menu */
-.navbar-links-mobile {
-  display:table-column;
-  background: rgb(0, 0, 0);
-  z-index:-1;
-  position: fixed;
-  top: 70px;
-  right:400px;
-  opacity: 1;
-  padding: 0px 20px;
-  text-align: left;
-  gap: 20px;
-  transform: translateX(100%);
-  transition: transform 0.3s ease-in-out;
-  height: 100vh;
-}
-
-.navbar-links-mobile li {
-  list-style-type: none;
-}
-
-.navbar-links-mobile a {
-  font-family: serif;
-  font-size: 18px;
-  color: #ebe2e2;
-  text-decoration: none;
-  padding: 10px 40px;
-  display: block;
-  transition: 0.3s ease;
-  text-transform: uppercase;
-  opacity: 1;
-}
-
-.navbar-links-mobile a:hover {
-  color: #0ff;
-  transform: scale(1.1);
-}
-
-.navbar-links-mobile.active {
-  transform: translateX(0);
-}
-
-/* Media Query for Mobile */
-@media screen and (max-width: 768px) {
-  .navbar-links {
-    display: none;
-  }
-
-  .mobile-menu-icon {
-    display: block;
-  }
-
-  .navbar-links-mobile.active {
-    display: block;
-  }
-}
-
-.navbar-links-mobile button {
-  background-color:rgb(96, 43, 10);
-  border: 2px solid #00ffcc; /* Futuristic neon border */
-  color: #00ffcc;
-  font-size: 1rem;
-  
-  border-radius: 10px;
-  cursor: pointer;
-  position: relative;
-  z-index: 1;
-  transition: all 0.4s ease;
-  overflow: hidden;
-  padding: 8px 22px;
-  margin-left: 40px;
-}
-
-.navbar-links-mobile li button::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: -100%;
-  width: 100%;
-  height: 100%;
-  background: linear-gradient(120deg, rgba(0, 255, 204, 0.3), rgba(0, 255, 204, 0.1));
-  transition: 0.5s ease;
-  z-index: -1;
-}
-
-.navbar-links-mobile li button:hover::before {
-  left: 100%;
-}
-
-.navbar-links-mobile li button:hover {
-  color: #fff;
-  border-color: #fff;
-}
-
-.navbar-links-mobile li button:focus {
-  outline: none;
-}
-
-.navbar-links-mobile li button:active {
-  transform: scale(0.95);
-}
-
-
-.navbar-links a.active,
-
-.navbar-links-mobile a.active {
-
-  color: #0ff; /* Highlighted color */
-
-  text-shadow: 0 0 10px rgba(0, 255, 255, 1), 0 0 20px rgba(0, 255, 255, 0.8);
-
-  transform: translateY(-2px);
-
-}
-
-/* Additional futuristic glow effect */
-.navbar-links-mobile li button::after {
-  content: "";
-  position: absolute;
-  top: -5px;
-  left: -5px;
-  right: -5px;
-  bottom: -5px;
-  background: rgba(0, 255, 204, 0.2);
-  border-radius: 5px;
-  z-index: -1;
-  opacity: 0;
-  transition: opacity 0.4s ease;
-}
-
-.navbar-links-mobile li button:hover::after {
-  opacity: 1;
-}
